Add files via upload
A few sample designs from recent projects:

Community Legal Advisors - Two Flyer designs I created for a presentation about the important of having a will called "The Elephant In The Room"

FairShare PA Wire Frames - These are the mock-up wire frames I created for the Fair Share PA landing page. You can see the landing page at demo.fairsharepa.org.

Good Law | Bad Law - The first two designs are social media posts to celebrate our 100th episode, I created all of the elements in each design. The third design is a rack card design I created for an expo we were attending. The background of the card is all of the keywords from our episodes at that time.

WeTrain - The App Screens folder contains images I created in Photoshop and Illustrator mapping each of the main app screens for the client app. Each screen is built off of a wire frame I designed and our team approved before moving on to this step. These designs were used to show the development team exactly how we wanted the app to look. In addition, my partner and I created an in depth work flow for every function of the app to accompany these designs for the development team 
The Coffee Shop flyer is a design I made to hand out at weekly meetups.
Laura Header, is an email header I designed for one of our weekly newsletters.
Next, you will see the business cards I designed for each of the founding partners and our head trainer.
Finally, is a social media ad I created.

YNPN - The first two images are two designs I created for a meetup the organization was arranging at a local coffee shop and the third image is a version of the logo I created, this was used on all of the social media platforms.